% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitstream-vera.R
\name{font_bitstream_vera}
\alias{font_bitstream_vera}
\title{Bitstream Vera fonts}
\usage{
font_bitstream_vera(variant = "sans", style = "roman", ext = "ttf")
}
\arguments{
\item{variant}{Variant of the font, e.g. \code{"sans"} or \code{"serif"}.}

\item{style}{Style of the font, e.g. \code{"roman"} or \code{"bold"}.}

\item{ext}{One of \code{"ttf"}, \code{"svg"} or \code{"woff"}.}
}
\value{
Path to the requested font. The version number of the font
  is given in the \code{version} attribute.
}
\description{
Available faces: \code{sans} (\code{roman}, \code{bold},
\code{oblique}, \code{oblique-bold}), \code{sans-mono}
(\code{roman}, \code{bold}, \code{oblique}, \code{bold-oblique}),
\code{serif} (\code{roman}, \code{bold}).
}
\references{
\url{https://www.gnome.org/fonts/}
}

