% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-dejavu.R
\name{font_dejavu}
\alias{font_dejavu}
\title{Dejavu fonts}
\usage{
font_dejavu(variant = "sans", style = "book", ext = "ttf")
}
\arguments{
\item{variant}{Variant of the font, e.g. \code{"sans"} or \code{"serif"}.}

\item{style}{Style of the font, e.g. \code{"roman"} or \code{"bold"}.}

\item{ext}{Either \code{"ttf"} or \code{"woff"}.}
}
\value{
The path to the requested font is provided in the
  \code{file} field. The version number of the font is given in the
  \code{version} field.
}
\description{
Available variants (style): \code{sans} (\code{book}, \code{bold},
\code{oblique}, \code{bold-oblique}, \code{extra-light}),
\code{sans-condensed} (\code{book}, \code{bold}, \code{oblique},
\code{bold-oblique}), \code{sans-mono} (\code{book}, \code{bold},
\code{oblique}, \code{bold-oblique}), \code{serif} (\code{book},
\code{bold}, \code{italic}, \code{bold-italic}),
\code{serif-condensed} (\code{book}, \code{bold}, \code{italic},
\code{bold-italic}), \code{math-tex-gyre} (\code{regular}).
}
\references{
\url{http://dejavu-fonts.org}
}

