% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_desc.R
\docType{data}
\name{font_desc}
\alias{font_desc}
\alias{font_desc_bitstream_vera}
\alias{font_desc_dejavu}
\alias{font_desc_liberation}
\title{font description to be used by graphical devices}
\format{an object that describes a font}
\usage{
font_desc

font_desc_bitstream_vera

font_desc_dejavu

font_desc_liberation
}
\description{
Summarize a font package.
}
\details{
\code{font_desc} is only there to provide common structure and behavior
for a given font description.

\code{font_desc_bitstream_vera} return details about Bitstream Vera
font families and styles.

\code{font_desc_dejavu} return details about DejaVu
font families and styles.

\code{font_desc_dejavu} return details about Liberation
font families and styles.
}
\section{Fields}{

\describe{
\item{\code{get_data}}{get a data.frame describing a font}

\item{\code{get_id}}{get font id}

\item{\code{get_families}}{get a list to be used by a graphic device.}

\item{\code{set_data,set_id}}{convenience function to be used when creating a
font description object.}
}}
\examples{
# construct a \\code{font_desc} object
font_desc_ <- font_desc_bitstream_vera$new()
font_desc_$get_families()
font_desc_$get_data()

font_desc_dejavu$new()$get_families()
font_desc_dejavu$new()$get_data()
font_desc_liberation$new()$get_families()
font_desc_liberation$new()$get_data()
}
\keyword{datasets}

