% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_desc.R
\name{font_desc_html_dependency}
\alias{font_desc_html_dependency}
\title{include fonts as as CSS dependency}
\usage{
font_desc_html_dependency(font_description = font_desc_bitstream_vera$new())
}
\arguments{
\item{font_description}{a \code{font_desc} object}
}
\description{
include fonts for families \code{sans}, \code{serif}, \code{mono}
as as CSS dependency.
}

