% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{font_faces}
\alias{font_faces}
\title{Font faces for a font family}
\usage{
font_faces(fontset, family)
}
\arguments{
\item{fontset}{A string giving the name of a set of fonts
(e.g. \code{"Bitstream Vera"}).}

\item{family}{One of \code{"sans"}, \code{"serif"}, \code{"mono"}
or \code{"symbol"}.}
}
\description{
Font faces for a font family
}

