% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{font_families}
\alias{font_families}
\title{Font families}
\usage{
font_families(fontset)
}
\arguments{
\item{fontset}{A string giving the name of a set of fonts
(e.g. \code{"Bitstream Vera"}).}
}
\description{
Returns families of font files for a given font set. The files are
organised by R nomenclature of families (\code{"sans"},
\code{"serif"}, \code{"mono"}, and \code{"symbol"}) and faces
(\code{"plain"}, \code{"italic"}, \code{"bold"}, and
\code{"bolditalic"}). When a font does not have a combination,
\code{NA} is reported.
}
\examples{
font_families("Bitstream Vera")$mono$bold
}

