% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{font_family}
\alias{font_family}
\title{Font family}
\usage{
font_family(fontset, family)
}
\arguments{
\item{fontset}{A string giving the name of a set of fonts
(e.g. \code{"Bitstream Vera"}).}

\item{family}{One of \code{"sans"}, \code{"serif"}, \code{"mono"}
or \code{"symbol"}.}
}
\description{
Font family
}

