% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{font_info}
\alias{font_info}
\title{Font metadata}
\usage{
font_info(font)
}
\arguments{
\item{font}{A string giving the name of the font
(e.g. \code{"bitstream_vera"}).}
}
\description{
This returns a data frame containing metadata about fonts. The
columns "variant" and "style" correspond to the properties of the
font as reported by fontconfig. These are suitable as parameters of
a font getter function such as \code{\link{font_bitstream_vera}()}.
The columns "family", "italic" and "bold" provide information in
terms of R nomenclature of fonts. The column "base" gives the base
filename of the font.
}
\seealso{
\code{\link{font_info_files}()}
}

