% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{font_info_files}
\alias{font_info_files}
\title{Font files}
\usage{
font_info_files(font)
}
\arguments{
\item{font}{A string giving the name of the font
(e.g. \code{"bitstream_vera"}).}
}
\description{
Returns a list of font files. The files are organised by R
nomenclature of families (\code{"sans"}, \code{"serif"},
\code{"mono"}, and \code{"symbol"}) and fonts (\code{"plain"},
\code{"italic"}, \code{"bold"}, and \code{"bolditalic"}). When a
font does not have a combination, \code{NA} is reported.
}
\examples{
font_info_files("bitstream_vera")$mono$bold
}

