% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-install.R
\name{font_install}
\alias{font_install}
\title{Download and install a font}
\usage{
font_install(font = "dejavu")
}
\arguments{
\item{font}{Only \code{dejavu} is currently available for download.}
}
\description{
Fonts that are too large to be stored on CRAN can be downloaded
manually. They are stored along other fonts in in the library path
under \code{fontquiver/fonts}.
}

