% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{font_register}
\alias{font_register}
\title{Register a font to fontquiver}
\usage{
font_register(fontset, getter, files)
}
\arguments{
\item{fontset}{Name of the font set.}

\item{getter}{Constructor of \code{font_file} objects.}

\item{files}{Data structure containing font information.}
}
\description{
Register a font to fontquiver
}

