% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{font_styles}
\alias{font_styles}
\title{Font styles for a font variant}
\usage{
font_styles(fontset, variant)
}
\arguments{
\item{fontset}{A string giving the name of a set of fonts
(e.g. \code{"Bitstream Vera"}).}

\item{variant}{Font variant, as per Fontconfig's nomenclature.}
}
\description{
Font styles for a font variant
}

