% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{font_variant}
\alias{font_variant}
\title{Font variant}
\usage{
font_variant(fontset, variant)
}
\arguments{
\item{fontset}{A string giving the name of a set of fonts
(e.g. \code{"Bitstream Vera"}).}

\item{variant}{Font variant, according to Fontconfig's
nomenclature.}
}
\description{
Font variant
}

