% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{font_variants}
\alias{font_variants}
\title{Font variants}
\usage{
font_variants(fontset)
}
\arguments{
\item{fontset}{A string giving the name of a set of fonts
(e.g. \code{"Bitstream Vera"}).}
}
\description{
Font variants
}

