% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splice_fonts}
\alias{splice_fonts}
\title{Splice fonts and font collections}
\usage{
splice_fonts(...)
}
\arguments{
\item{...}{Fonts or collections of fonts. See
\code{\link{font_families}()} and \code{\link{font_variants}()}
for creating collections of fonts. You can also supply lists of
individual fonts as returned by \code{\link{fonts}()}.}
}
\description{
\code{splice_fonts()} Reduces its arguments to a flat list. It
accepts indinstinctly \code{font} objects, lists of \code{font}
objects (obtained with \code{\link{fonts}()}), or collections of
fonts produced by \code{\link{font_variants}()} or
\code{\link{font_families}()}. Duplicate fonts are removed from the
result.
}
\examples{
splice_fonts(font("Bitstream Vera", "Sans", "Oblique"), font_faces("Bitstream Vera", "mono"))
}

