% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{font_info}
\alias{font_info}
\title{Font metadata}
\usage{
font_info(fontset)
}
\arguments{
\item{fontset}{A string giving the name of a set of fonts
(e.g. \code{"Bitstream Vera"}).}
}
\description{
This returns a data frame containing metadata about fonts. The
columns "variant" and "style" correspond to the properties of the
font as reported by fontconfig. These are suitable as parameters of
a font getter function such as \code{\link{font_bitstream_vera}()}.
The columns "family", "italic" and "bold" provide information in
terms of R nomenclature of fonts. The column "base" gives the base
filename of the font.
}
\seealso{
\code{\link{font_families}()}, \code{\link{font_variants}()}
}

