% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontset.R
\name{fontset_register}
\alias{fontset_register}
\title{Register a font to fontquiver}
\usage{
fontset_register(fontset, getter, files)
}
\arguments{
\item{fontset}{Name of the font set.}

\item{getter}{Constructor of \code{font} objects.}

\item{files}{Data structure containing font information.}
}
\description{
Register a font to fontquiver
}

