% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-dependency.R
\name{htmlFontDependency}
\alias{htmlFontDependency}
\title{Include font as CSS dependency}
\usage{
htmlFontDependency(font_getter = font_bitstream_vera, variant = NULL,
  style = NULL)
}
\arguments{
\item{font_getter}{A fontquiver getter function such as
\code{font_bitstream_vera}.}

\item{variant}{Variant of the font, e.g. \code{"sans"} or \code{"serif"}.}

\item{style}{Style of the font, e.g. \code{"roman"} or \code{"bold"}.}
}
\description{
Include font as CSS dependency
}
\examples{
# Create an htmlDependency object:
dep <- htmlFontDependency(variant = "serif")

# Check the name of the embedded font
dep$name

# Use that name in your dependent css or html files. For example:
# body {
#  font-family: fontquiver-bitstream-vera-sans-roman, sans-serif;
# }
}

